Number_Guessing_Game.java
Number Guessing Game is a simple Java program that allows the user to guess a randomly generated number within a specified range. The program provides hints to the user if the guess number   is too high or too low and keeps track of the number of attempts made by the user.